New lab: bridge using arp proxy and dhcp relay
Bridging is sometimes hard (for example bridging wired and wireless
interfaces). So there is a network-lab to test a layer 3 bridge using
proxy_arp and dhcp relay.

R1 need a route to all clients behind eth1, so we split dhcp range
(192.168.1.0/24) into two subnets (192.168.1.0/25 and 192.168.1.128/25)
and detect relay on R0 using agent-circuit-id (which is a pretty bad
hack).

The lab:

              [C1]
               |
               |
          [switch S1]
               |
               | eth1
              [R1] dhcp relay
               | eth0
               |
[C0] <==> [switch S0] <==> [R0] dhcp server

Signed-off-by: Philippe Pepiot <phil@philpep.org>